x86: Drop task from thread_info, add accessor functions
Add accessor functions to get a thread_info from a task_struct, and
vice-versa.  This also allows the task_struct pointer to be dropped
from thread_info altogether.

This patch also adds some accessor functions to some of the thread_info
fields, mainly for ones that are derived from a task_struct to begin
with.

Signed-off-by: Peter P Waskiewicz Jr <peter.p.waskiewicz.jr@intel.com>